+++ android/app/src/main/java/kr/pibutenten/app/MainActivity.java
@@ -1,0 +1,5 @@
+package kr.pibutenten.app;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}